--- src/main/java/com/perfumepedia/PerfumePedia/domain/Discontinue.java
@@ -1,5 +1,0 @@
-package com.perfumepedia.PerfumePedia.domain;
-
-public enum Discontinue {
-    CONTINUE, DISCONTINUE
-}
