--- src/main/java/com/perfumepedia/PerfumePedia/domain/NoteType.java
@@ -1,5 +1,0 @@
-package com.perfumepedia.PerfumePedia.domain;
-
-public enum NoteType {
-    TOP, MIDDLE, BASE, SINGLE
-}
