--- src/main/java/com/perfumepedia/PerfumePedia/domain/WordType.java
@@ -1,5 +1,0 @@
-package com.perfumepedia.PerfumePedia.domain;
-
-public enum WordType {
-    BRAND, PERFUME, NOTE
-}
