+++ src/main/java/com/perfumepedia/PerfumePedia/domain/entity/Discontinue.java
@@ -1,0 +1,5 @@
+package com.perfumepedia.PerfumePedia.domain.entity;
+
+public enum Discontinue {
+    CONTINUE, DISCONTINUE
+}
