+++ src/main/java/com/perfumepedia/PerfumePedia/domain/entity/Image.java
@@ -1,0 +1,21 @@
+package com.perfumepedia.PerfumePedia.domain.entity;
+
+import jakarta.persistence.Embeddable;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Embeddable
+@Getter
+public class Image {
+
+    private String name;
+    private String path;
+
+    // 생성자
+    public Image(String name, String path){
+        this.name = name;
+        this.path = path;
+    }
+}
