+++ src/main/java/com/perfumepedia/PerfumePedia/domain/entity/NoteType.java
@@ -1,0 +1,5 @@
+package com.perfumepedia.PerfumePedia.domain.entity;
+
+public enum NoteType {
+    TOP, MIDDLE, BASE, SINGLE
+}
