+++ src/main/java/com/perfumepedia/PerfumePedia/domain/entity/WordType.java
@@ -1,0 +1,5 @@
+package com.perfumepedia.PerfumePedia.domain.entity;
+
+public enum WordType {
+    BRAND, PERFUME, NOTE
+}
