+++ src/main/java/com/perfumepedia/PerfumePedia/domain/service/PerfumeDetailService.java
@@ -1,0 +1,4 @@
+package com.perfumepedia.PerfumePedia.domain.service;
+
+public class PerfumeDetailService {
+}
