--- src/main/java/com/perfumepedia/PerfumePedia/dto/ResponseData.java
@@ -1,11 +1,0 @@
-package com.perfumepedia.PerfumePedia.dto;
-
-public record ResponseData<T>(T data) {
-    /**
-     * 생성 및 초기화 진행
-     *
-     * @param data T 객체
-     */
-    public ResponseData {
-    }
-}
