--- src/main/java/com/perfumepedia/PerfumePedia/service/PerfumeDetailService.java
@@ -1,4 +1,0 @@
-package com.perfumepedia.PerfumePedia.service;
-
-public class PerfumeDetailService {
-}
